Project-1/requirements.txt
omegaconf
lpips
numpy
networkx
pandas
scikit-image
scikit-learn
imageio
scipy
fsspec
matplotlib
seaborn
plotly
psycopg-binary
sqlalchemy
h5py
albumentations
skorch
torchmetrics
torchtext
transformers
tensorboard
timm
captum
pyarrow
fastparquet
torchaudio
pytorch-lightning
minio
coloredlogs
GitPython
pickle5
bertviz
webdataset
dvc
pyspark
shap
ipyparallel
graphviz
missingno
python-dotenv
einops
opt_einsum
ffmpeg
cython 
moviepy 
psutil 
scipy  
ipykernel
thop
imread-from-url
gdown
more-itertools
streamlit
DALL-E
https://github.com/podgorskiy/dnnlib/releases/download/0.0.1/dnnlib-0.0.1-py3-none-any.whl
